Added dynamic data acces
In this version i added a way to access user data via the connection string which is store together with the username and project number in the database. The database for the projects has also been updated to store navbar and chart data.

Nextup i will need to create a way to show the data, i need to build an easy user experience for generating an SQL query.