Fixed a lot of small things, and added a new JS library which does not work yet.
Small changes:

o Fixed the zooming functionality
-- To do this i added a function to lock the charts.

o Fixed the timeline which was not shown in order.

The library i added is hammerjs which should add touchscreen like user input. I added this library to get de panning functionality of the charts to work, but it does not.

Will search for a fix.